+++ Base/res/html/inspector/inspector.css
@@ -1,0 +1,169 @@
+body {
+    font-family: system-ui, sans-serif;
+    font-size: 10pt;
+
+    margin: 0;
+}
+
+.split-view {
+    width: 100vw;
+    height: 100vh;
+}
+
+.split-view-container {
+    overflow: scroll;
+}
+
+.tab-controls-container {
+    position: absolute;
+    width: 100%;
+
+    padding: 4px;
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    z-index: 100;
+}
+
+.tab-controls {
+    overflow: hidden;
+    flex-shrink: 0;
+}
+
+.tab-controls button {
+    font-size: 12px;
+    font-weight: 600;
+
+    float: left;
+    border: none;
+    outline: none;
+    cursor: pointer;
+
+    padding: 4px 8px;
+}
+
+.tab-controls :first-child {
+    border-radius: 0.5rem 0 0 0.5rem;
+}
+
+.tab-controls :last-child {
+    border-radius: 0 0.5rem 0.5rem 0;
+}
+
+.tab-content {
+    height: 100%;
+
+    display: none;
+    border-radius: 0.5rem;
+
+    margin-top: 30px;
+    padding: 8px;
+}
+
+@media (prefers-color-scheme: dark) {
+    html {
+        background-color: rgb(23, 23, 23);
+    }
+
+    .tab-controls-container {
+        background-color: rgb(57, 57, 57);
+    }
+
+    .tab-controls button {
+        color: white;
+        background-color: rgb(43, 42, 50);
+    }
+
+    .tab-controls button.active {
+        background-color: rgb(22 100 219);
+    }
+
+    .tab-controls button + button {
+        border-left: 1px solid rgb(96, 96, 96);
+    }
+}
+
+@media (prefers-color-scheme: light) {
+    .tab-controls-container {
+        background-color: rgb(229, 229, 229);
+    }
+
+    .tab-controls button {
+        color: black;
+        background-color: white;
+    }
+
+    .tab-controls button.active {
+        color: white;
+        background-color: rgb(28, 138, 255);
+    }
+
+    .tab-controls button + button {
+        border-left: 1px solid rgb(242, 242, 242);
+    }
+}
+
+details > :not(:first-child) {
+    display: list-item;
+    list-style: none inside;
+    margin-left: 1em;
+}
+
+.hoverable {
+    display: block;
+    padding: 1px;
+}
+
+@media (prefers-color-scheme: dark) {
+    .hoverable:hover {
+        background-color: #31383e;
+    }
+
+    .selected {
+        border: 1px dashed cyan;
+        padding: 0;
+    }
+}
+
+@media (prefers-color-scheme: light) {
+    .hoverable:hover {
+        background-color: rgb(236, 236, 236);
+    }
+
+    .selected {
+        border: 1px dashed blue;
+        padding: 0;
+    }
+}
+
+.property-table {
+    width: 100%;
+
+    table-layout: fixed;
+    border-collapse: collapse;
+}
+
+.property-table th {
+    position: sticky;
+    top: 0px;
+}
+
+.property-table th,
+.property-table td {
+    padding: 4px;
+    text-align: left;
+}
+
+@media (prefers-color-scheme: dark) {
+    .property-table th {
+        background-color: rgb(57, 57, 57);
+    }
+}
+
+@media (prefers-color-scheme: light) {
+    .property-table th {
+        background-color: rgb(229, 229, 229);
+    }
+}
